requirements.txt
vllm>=0.4.0
transformers
unbabel-comet
